record/src/requirements.txt
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mutagen==1.46.0
psycopg2-binary==2.9.6
pycryptodomex==3.16.0
pytube==12.1.3
requests==2.28.2
urllib3==1.26.15
websockets==10.4
Werkzeug==2.2.2
yt-dlp==2023.3.4
